Declarations in src
- moved all delcarations into the src directory
- split each entity interface and associated types/interfaces into
individual declaration files
- removed "module" declaration; all declaration files now simply export
the types/interfaces
