+++ src/index.d.ts
@@ -1,0 +1,78 @@
+import {Alert} from './alert'
+import {AtlasUser} from './atlasUser';
+import {Cluster} from './cluster';
+import {CustomDbRole} from './customDbRole';
+import {Event} from './event';
+import {Organization} from './organization';
+import {Project} from './project';
+import {ProjectWhitelist} from './projectWhitelist';
+import {User} from './user';
+
+export interface KeyValuePairDocument {
+    key: string;
+    value: string;
+}
+
+export type KeyValuePairDocumentArray = KeyValuePairDocument[];
+
+export interface AtlasResultsResponse<T> {
+    results: T[];
+    links: Links;
+    totalCount: number;
+}
+
+export interface Link {
+    href: string;
+    rel: string;
+}
+
+export type Links = Link[];
+
+export interface AtlasError {
+    details: string;
+    error: number;
+    errorCode: string;
+    reason: string;
+}
+
+export type ResponseOrError<T> = T | AtlasError;
+
+// Atlas Client
+export interface AtlasClient {
+    user: User;
+    alert: Alert;
+    atlasUser: AtlasUser;
+    organization: Organization;
+    project: Project;
+    projectWhitelist: ProjectWhitelist;
+    customDbRole: CustomDbRole;
+    cluster: Cluster;
+    event: Event;
+}
+
+export interface AtlasClientConfig {
+    /**
+     * API Access Public Key
+     */
+    publicKey: string;
+    /**
+     * API Access Private Key
+     */
+    privateKey: string;
+    /**
+     * Base URL for Atlas API
+     */
+    baseUrl: string;
+    /**
+     * Target Project ID in Atlas account
+     */
+    projectId: String;
+}
+
+export interface AtlasClientOptions {
+    envelope?: boolean;
+    itemsPerPage?: number;
+    pretty?: boolean;
+}
+
+export default function getMongodbAtlasApiClient(config: AtlasClientConfig): AtlasClient;
